First Push
-Don't know what to gitignore yet
-Unity Sucks
-Finally got the stupid stat box to work
Couple of Ideas: Use image of char sheet as one scene, add the ability to interact, then add another scene to deal with fighting, perhaps another scene to deal with a journal, etc.
Alternatively, I could just place it all in the same scene and shift the focus of the camera, but that sounds like a pain in the butt. Both ways are probably simple but my ignorance of Unity is very annoying